intégration with primeNg (profil, admin, cm, formateur, apprenant, ref)

+++ frontFilRouge/src/app/acceuil/admin/admin-user/list-admin/list-admin.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, RouterStateSnapshot } from '@angular/router';
+import { ConfirmationService } from 'primeng/api';
 import { AdminUserService } from 'src/app/service/adminUserService/admin-user.service';
 
 @Component({
@@ -17,10 +18,14 @@ export class ListAdminComponent implements OnInit {
   snapshot: RouterStateSnapshot;
   id = [];
   url: string;
+  first = 0;
+  rows = 5;
+  clonedProducts: { } = {};
 
   constructor(
     private adminUserService: AdminUserService,
-    private router: Router
+    private router: Router,
+    private confirmationService: ConfirmationService
   ) { }
 
   ngOnInit(): void {
@@ -62,24 +67,51 @@ export class ListAdminComponent implements OnInit {
     return false;
   }
 
-  onClickBtnPut(id) {
+
+  onRowEditInit(id) {
     let currentUrl = this.router.url;
     this.router.routeReuseStrategy.shouldReuseRoute = () => false;
     this.router.onSameUrlNavigation = 'reload';
     return this.router.navigate([`/acceuil/liste/admins/${id}/modifier`]);
   }
 
-  onClickBtnDelete() {
-    var toDelete: number;
-    toDelete = this.getIdOnUrl();
-    console.log(toDelete);
-    //var toRemove = this.profils.slice().pop();
-     return this.adminUserService.deleteAdmin(toDelete).subscribe(
+  onRowEditSave(admin, id : number) {
+    return this.adminUserService.putAdmin(id, admin).subscribe(
       (res: any) => { 
-        this.reloadComponent();
         console.log(res)
       }
     );
+  }
+
+  onRowEditCancel(admin, index: number) {
+    this.admins[index] = this.clonedProducts[admin.id];
+    delete this.clonedProducts[admin.id];
+  }
+
+  confirm(event: Event) {
+    this.confirmationService.confirm({
+        target: event.target,
+        message: 'Êtes-vous sure de vouloir supprimer cet Admin?',
+        icon: 'pi pi-exclamation-triangle',
+        accept: () => {
+          var toDelete: number;
+          toDelete = this.getIdOnUrl();
+          console.log(toDelete);
+          //var toRemove = s.slice().pop();
+           return this.adminUserService.deleteAdmin(toDelete).subscribe(
+            (res: any) => { 
+              this.reloadComponent();
+              console.log(res)
+            }
+          );
+        },
+        reject: () => {
+          // return this.reloadComponent();
+        }
+    });
+  }
+
+  onClickBtnDelete() {
   }
 
 }
